+++ libs/pages/dashboard/src/lib/pages-dashboard.tsx
@@ -1,6 +1,5 @@
 import { Fragment, useEffect } from 'react';
 import DashboardHeader from './dashboard-header/dashboard-header';
-import DashboardOverview from './dashboard-overview/dashboard-overview';
 import DashboardOrders from './dashboard-orders/dashboard-orders';
 import { useDashboardFacade } from './dashboard-store/dashboard.facade';
 
@@ -19,7 +18,6 @@ export function PagesDashboard() {
       <main className="flex-grow flex-1">
         <div className="mt-8">
           <div className="max-w-6xl mx-auto px-4 sm:px-6 lg:px-8">
-            <DashboardOverview />
             <DashboardOrders />
           </div>
         </div>
